requirements.txt
dj-rest-auth==4.0.1
Django==4.2.3
django-allauth==0.54.0
djangorestframework==3.14.0
drf-yasg==1.21.7
Pillow==10.0.0
PyYAML==6.0.1
